Fix unpacking of DDC type bits
DDC type is a 1-, 2-, or 3-bit value depending on the revision
of the loaded bitstream. When the second bit was added, it was added
as the high-order bit, which left existing type codes unchanged.
However, the third bit was added as the low-order bit, which shifted
the existing type codes. Unpacking is now done using three arrays
which map from 1-, 2-, or 3-bit values to the correct associated
DDC types.

A new type, DDCUNDEFINED, has been added and is returned in cases
where the DDC type code cannot be decoded. Previously, DD4DECIMATE
was returned where decoding failed.